bot/requirements.txt
aiogram>=2.17.1
aiohttp>=3.8.1
googlesearch-python>=1.0.1
sqlmodel>=0.0.5